# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Homepage Backend API",
    "description": "A flexible content management API for handling various entry types with validation and notifications",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.chenmo1212.cn/message",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/api/v1/message/entries": {
      "get": {
        "tags": ["Entries"],
        "summary": "Get entries list",
        "description": "Retrieve a paginated list of entries with optional filters",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Filter by entry type",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Filter by source",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_show",
            "in": "query",
            "description": "Filter by visibility status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "description": "Filter by tags (can be multiple)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Entry"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "tags": ["Entries"],
        "summary": "Create new entry",
        "description": "Create a new entry with validation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEntryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entry created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "example": "Entry created successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        }
                      }
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/message/entries/batch-delete": {
      "post": {
        "tags": ["Entries"],
        "summary": "Batch delete entries",
        "description": "Delete multiple entries at once",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id_list"],
                "properties": {
                  "id_list": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": ["507f1f77bcf86cd799439011", "507f1f77bcf86cd799439012"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entries deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "example": "Deleted 2 entries"
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/message/admin/entries": {
      "get": {
        "tags": ["Admin"],
        "summary": "Admin get all entries",
        "description": "Retrieve all entries including hidden and deleted ones",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Filter by entry type",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Filter by source",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AdminEntry"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/message/admin/entries/{entry_id}/status": {
      "put": {
        "tags": ["Admin"],
        "summary": "Update entry status",
        "description": "Update the status fields of an entry",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "description": "Entry ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "is_show": {
                    "type": "boolean"
                  },
                  "is_delete": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "example": "Entry status updated successfully"
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/message/admin/entries/stats": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get statistics",
        "description": "Get entry statistics with optional filters",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Filter by entry type",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Filter by source",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "description": "Statistics data"
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/message/admin/types": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get all supported types",
        "description": "Retrieve all configured entry types",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TypeInfo"
                      }
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Create new type",
        "description": "Add a new entry type configuration",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type", "config"],
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "blog_post"
                  },
                  "config": {
                    "type": "object",
                    "description": "Type configuration with schema"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Type created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "example": "Type blog_post created successfully"
                    },
                    "status": {
                      "type": "integer",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Entry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "type": {
            "type": "string",
            "example": "message"
          },
          "source": {
            "type": "string",
            "example": "homepage"
          },
          "metadata": {
            "type": "object",
            "description": "Entry-specific data based on type"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "timestamps": {
            "$ref": "#/components/schemas/Timestamps"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["important", "urgent"]
          }
        }
      },
      "AdminEntry": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Entry"
          },
          {
            "type": "object",
            "properties": {
              "agent": {
                "type": "string",
                "description": "User agent string"
              }
            }
          }
        ]
      },
      "Status": {
        "type": "object",
        "properties": {
          "is_show": {
            "type": "boolean",
            "example": true
          },
          "is_delete": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "Timestamps": {
        "type": "object",
        "properties": {
          "create_time": {
            "type": "string",
            "format": "date-time"
          },
          "update_time": {
            "type": "string",
            "format": "date-time"
          },
          "admin_time": {
            "type": "string",
            "format": "date-time"
          },
          "delete_time": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "total": {
            "type": "integer",
            "example": 100
          },
          "pages": {
            "type": "integer",
            "example": 5
          }
        }
      },
      "CreateEntryRequest": {
        "type": "object",
        "required": ["type", "metadata"],
        "properties": {
          "type": {
            "type": "string",
            "example": "message"
          },
          "metadata": {
            "type": "object",
            "description": "Entry-specific data"
          },
          "source": {
            "type": "string",
            "example": "homepage"
          },
          "agent": {
            "type": "string",
            "description": "User agent string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "TypeInfo": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "message"
          },
          "name": {
            "type": "string",
            "example": "Message"
          },
          "description": {
            "type": "string",
            "example": "User messages from homepage"
          },
          "notification_enabled": {
            "type": "boolean",
            "example": true
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error message"
                },
                "status": {
                  "type": "integer",
                  "example": 400
                }
              }
            }
          }
        }
      }
    }
  }
}